requirements.txt
anyio==4.3.0
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.1.1
python-dotenv==1.0.1
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
